Lambdas/Test/requirements.txt
selenium==3.14.0
chromedriver-binary==2.37.0
lxml == 4.6.1
pymongo == 3.11.3
python_dotenv == 0.17.0
requests == 2.24.0
selenium_wire == 4.0.4